Fix dragging custom mime types from tree view into an editor
Fixes #175589

Introduces a new `AsyncDataTransferStore` class to facilitate setting data transfer values on a drag/copy event async
