+++ src/vs/editor/contrib/dropIntoEditor/browser/dropIntoEditorContribution.ts
@@ -12,21 +12,24 @@ import { addExternalEditorsDropData, toVSDataTransfer } from 'vs/editor/browser/
 import { ICodeEditor } from 'vs/editor/browser/editorBrowser';
 import { EditorContributionInstantiation, registerEditorContribution } from 'vs/editor/browser/editorExtensions';
 import { IBulkEditService, ResourceTextEdit } from 'vs/editor/browser/services/bulkEditService';
+import { EditorOption } from 'vs/editor/common/config/editorOptions';
 import { IPosition } from 'vs/editor/common/core/position';
 import { Range } from 'vs/editor/common/core/range';
 import { IEditorContribution } from 'vs/editor/common/editorCommon';
 import { DocumentOnDropEdit, WorkspaceEdit } from 'vs/editor/common/languages';
 import { TrackedRangeStickiness } from 'vs/editor/common/model';
 import { ILanguageFeaturesService } from 'vs/editor/common/services/languageFeatures';
+import { DraggedTreeItemsIdentifier } from 'vs/editor/common/services/treeViewsDnd';
+import { ITreeViewsDnDService } from 'vs/editor/common/services/treeViewsDndService';
 import { PostDropWidgetManager } from 'vs/editor/contrib/dropIntoEditor/browser/postDropWidget';
 import { CodeEditorStateFlag, EditorStateCancellationTokenSource } from 'vs/editor/contrib/editorState/browser/editorState';
+import { InlineProgressManager } from 'vs/editor/contrib/inlineProgress/browser/inlineProgress';
 import { SnippetParser } from 'vs/editor/contrib/snippet/browser/snippetParser';
+import { localize } from 'vs/nls';
+import { LocalSelectionTransfer } from 'vs/platform/dnd/browser/dnd';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { IWorkspaceContextService } from 'vs/platform/workspace/common/workspace';
 import { registerDefaultDropProviders } from './defaultOnDropProviders';
-import { EditorOption } from 'vs/editor/common/config/editorOptions';
-import { InlineProgressManager } from 'vs/editor/contrib/inlineProgress/browser/inlineProgress';
-import { localize } from 'vs/nls';
 
 
 export class DropIntoEditorController extends Disposable implements IEditorContribution {
@@ -39,12 +42,15 @@ export class DropIntoEditorController extends Disposable implements IEditorContr
 	private readonly _dropProgressManager: InlineProgressManager;
 	private readonly _postDropWidgetManager: PostDropWidgetManager;
 
+	private readonly treeItemsTransfer = LocalSelectionTransfer.getInstance<DraggedTreeItemsIdentifier>();
+
 	constructor(
 		editor: ICodeEditor,
-		@IBulkEditService private readonly _bulkEditService: IBulkEditService,
-		@ILanguageFeaturesService private readonly _languageFeaturesService: ILanguageFeaturesService,
 		@IInstantiationService instantiationService: IInstantiationService,
 		@IWorkspaceContextService workspaceContextService: IWorkspaceContextService,
+		@IBulkEditService private readonly _bulkEditService: IBulkEditService,
+		@ILanguageFeaturesService private readonly _languageFeaturesService: ILanguageFeaturesService,
+		@ITreeViewsDnDService private readonly _treeViewsDragAndDropService: ITreeViewsDnDService,
 	) {
 		super();
 
@@ -126,9 +132,24 @@ export class DropIntoEditorController extends Disposable implements IEditorContr
 			return new VSDataTransfer();
 		}
 
-		const textEditorDataTransfer = toVSDataTransfer(dragEvent.dataTransfer);
-		addExternalEditorsDropData(textEditorDataTransfer, dragEvent);
-		return textEditorDataTransfer;
+		const dataTransfer = toVSDataTransfer(dragEvent.dataTransfer);
+		addExternalEditorsDropData(dataTransfer, dragEvent);
+
+		if (this.treeItemsTransfer.hasData(DraggedTreeItemsIdentifier.prototype)) {
+			const data = this.treeItemsTransfer.getData(DraggedTreeItemsIdentifier.prototype);
+			if (Array.isArray(data)) {
+				for (const id of data) {
+					const treeDataTransfer = await this._treeViewsDragAndDropService.removeDragOperationTransfer(id.identifier);
+					if (treeDataTransfer) {
+						for (const [type, value] of treeDataTransfer.entries()) {
+							dataTransfer.replace(type, value);
+						}
+					}
+				}
+			}
+		}
+
+		return dataTransfer;
 	}
 
 	private async applyDropResult(editor: ICodeEditor, position: IPosition, selectedEditIndex: number, allEdits: readonly DocumentOnDropEdit[], token: CancellationToken): Promise<void> {
